step 15: final solution

+++ counter/src/counter.cairo
@@ -16,24 +16,36 @@ trait IKillSwitch<TContractState> {
 mod Counter {
     use starknet::{ContractAddress};
     use super::{ICounter};
-    use kill_switch::{IKillSwitchDispatcher, IKillSwitchDispatcherTrait}; // Import the kill switch interface
+    use kill_switch::{IKillSwitchDispatcher, IKillSwitchDispatcherTrait};
+    use openzeppelin::access::ownable::OwnableComponent; // Import the kill switch interface
+
+    component!(path: OwnableComponent, storage: ownable, event: OwnableEvent);
+
+    #[abi(embed_v0)]
+    impl OwnableImpl = OwnableComponent::OwnableImpl<ContractState>;
+
+    impl OwnableInternalImpl = OwnableComponent::InternalImpl<ContractState>;
 
     #[storage]
-    struct Storage {
+    struct Storage {    
         counter: u32,
         kill_switch: ContractAddress, //Add a field for the kill switch contract
+        #[substorage(v0)]
+        ownable: OwnableComponent::Storage
     }
 
     #[constructor]
     fn constructor(ref self: ContractState, initial_counter: u32, kill_switch: ContractAddress) {
         self.counter.write(initial_counter);
         self.kill_switch.write(kill_switch); //Initialize the kill switch contract
+        self.ownable.initializer(initial_owner);
     }
 
     #[event]
     #[derive(Drop, starknet::Event)]
     enum Event {
         CounterIncreased: CounterIncreased,
+        OwnableEvent: OwnableComponent::Event,
     }
 
     #[derive(Drop, starknet::Event)]
@@ -49,6 +61,7 @@ mod Counter {
         }
 
         fn increase_counter(ref self: ContractState) {
+            self.ownable.assert_only_owner();
             let is_active = self.kill_switch.read().is_active(); //check if the contract is active
 
             if (IKillSwitchDispatcher { contract_address: self.kill_switch.read()}).is_active() { //Check if the kill switch is active
